fixed bug in slope function
