+++ test/ncl_tests/test_satvpr_slope_fao56.ncl
@@ -1,6 +1,6 @@
 temp = ispan(1, 100, 1)
 
-slope = satvpr_tdew_fao56(temp, (/2,2/))
+slope = satvpr_slope_fao56(temp, (/2,2/))
 
 system("/bin/rm -f satvpr_slope_fao56_output.nc")   ; remove any pre-existing file
 ncdf = addfile("satvpr_slope_fao56_output.nc" ,"c")  ; open output netCDF file
