--- test/test_max_daylight.py
@@ -1,84 +1,0 @@
-import sys
-import unittest
-
-import dask.array as da
-import dask.distributed as dd
-import numpy as np
-import xarray as xr
-
-# Import from directory structure if coverage test, or from installed
-# packages otherwise
-if "--cov" in str(sys.argv):
-    from src.geocat.comp import max_daylight
-else:
-    from geocat.comp import max_daylight
-
-
-class Test_max_daylight(unittest.TestCase):
-
-    @classmethod
-    def setUpClass(cls):
-
-        # get ground truth from ncl run netcdf file
-        try:
-            ncl_xr_gt = xr.open_dataarray(
-                "max_daylight_test.nc"
-            )  # Generated by running ncl_tests/test_max_daylight.ncl
-        except:
-            ncl_xr_gt = xr.open_dataarray("test/max_daylight_test.nc")
-
-        # set up ground truths
-        cls.ncl_gt = np.asarray(ncl_xr_gt)
-
-        cls.jday_gt = np.linspace(1, 365, num=365)
-        cls.lat_gt = np.linspace(-66, 66, num=133)
-
-        # make client to reference in subsequent tests
-        cls.client = dd.Client()
-
-    def test_numpy_input(self):
-        assert np.allclose(max_daylight(self.jday_gt, self.lat_gt),
-                           self.ncl_gt,
-                           atol=0.005)
-
-    def test_float_input(self):
-        assert np.allclose(max_daylight(246, -20.0), 11.66559, atol=0.005)
-
-    def test_list_input(self):
-        assert np.allclose(max_daylight(self.jday_gt.tolist(),
-                                        self.lat_gt.tolist()),
-                           self.ncl_gt,
-                           atol=0.005)
-
-    def test_xarray_input(self):
-        jday = xr.DataArray(self.jday_gt)
-        lat = xr.DataArray(self.lat_gt)
-
-        assert np.allclose(max_daylight(jday, lat), self.ncl_gt, atol=0.005)
-
-    def test_dask_unchunked_input(self):
-        jday = da.from_array(self.jday_gt)
-        lat = da.from_array(self.lat_gt)
-
-        out = self.client.submit(max_daylight, jday, lat).result()
-
-        assert np.allclose(out, self.ncl_gt, atol=0.005)
-
-    def test_dask_chunked_input(self):
-        jday = da.from_array(self.jday_gt, chunks='auto')
-        lat = da.from_array(self.lat_gt, chunks='auto')
-
-        out = self.client.submit(max_daylight, jday, lat).result()
-
-        assert np.allclose(out, self.ncl_gt, atol=0.005)
-
-    def test_input_dim(self):
-        self.assertRaises(ValueError, max_daylight,
-                          np.arange(4).reshape(2, 2),
-                          np.arange(4).reshape(2, 2))
-
-    def test_lat_bound_warning(self):
-        self.assertWarns(UserWarning, max_daylight, 10, 56)
-
-    def test_lat_bound_second_warning(self):
-        self.assertWarns(UserWarning, max_daylight, 10, 67)
